Move DecompressionCoordinator to :core
It's a special use case for a PersistentCache that is used via :core only.
