+++ platforms/core-execution/workers/src/main/java/org/gradle/workers/internal/WorkerDaemonServer.java
@@ -28,6 +28,8 @@ import org.gradle.api.internal.collections.DefaultDomainObjectCollectionFactory;
 import org.gradle.api.internal.collections.DomainObjectCollectionFactory;
 import org.gradle.api.internal.file.FileCollectionFactory;
 import org.gradle.api.internal.file.FileResolver;
+import org.gradle.api.internal.file.archive.DecompressionCoordinator;
+import org.gradle.api.internal.file.archive.DefaultDecompressionCoordinator;
 import org.gradle.api.internal.project.IsolatedAntBuilder;
 import org.gradle.api.internal.project.antbuilder.DefaultIsolatedAntBuilder;
 import org.gradle.api.internal.provider.DefaultProviderFactory;
@@ -38,8 +40,6 @@ import org.gradle.api.resources.ReadableResource;
 import org.gradle.api.resources.ResourceHandler;
 import org.gradle.api.resources.TextResourceFactory;
 import org.gradle.cache.UnscopedCacheBuilderFactory;
-import org.gradle.cache.internal.DecompressionCoordinator;
-import org.gradle.cache.internal.DefaultDecompressionCoordinator;
 import org.gradle.cache.internal.scopes.DefaultBuildTreeScopedCacheBuilderFactory;
 import org.gradle.cache.scopes.BuildTreeScopedCacheBuilderFactory;
 import org.gradle.initialization.LegacyTypesSupport;
