+++ subprojects/core/src/test/groovy/org/gradle/api/internal/file/archive/ZipFileTreeTest.java
@@ -20,7 +20,7 @@ import org.gradle.api.InvalidUserDataException;
 import org.gradle.api.file.EmptyFileVisitor;
 import org.gradle.api.internal.file.TestFiles;
 import org.gradle.api.internal.provider.Providers;
-import org.gradle.cache.internal.TestCaches;
+import org.gradle.cache.internal.TestDecompressionCoordinators;
 import org.gradle.test.fixtures.file.TestFile;
 import org.junit.Test;
 
@@ -45,7 +45,7 @@ public class ZipFileTreeTest extends AbstractArchiveFileTreeTest {
             fileSystem(),
             directoryFileTreeFactory(),
             fileHasher(),
-            TestCaches.decompressionCache(tempDirProvider.getTestDirectory().createDir("cache-dir")),
+            TestDecompressionCoordinators.decompressionCoordinator(tempDirProvider.getTestDirectory().createDir("cache-dir")),
             TestFiles.tmpDirTemporaryFileProvider(tempDirProvider.getTestDirectory()));
     }
